+++ src/parser/element.rs
@@ -135,7 +135,9 @@ fn parse_children(iter: &mut Iter<'_>, parent: &mut Element) -> Result<()> {
             }
         } else {
             let text = parse_text(iter)?;
-            parent.add_node(Node::Text(text));
+            if !text.is_empty() {
+                parent.add_node(Node::Text(text));
+            }
         }
         // some parsing functions may return with the iter pointing to the last thing that was part
         // of their construct, while others might advance the iter to the next char *after* the