+++ src/parser/string.rs
@@ -1,5 +1,6 @@
 use crate::error::Result;
 use crate::parser::Iter;
+use crate::xdoc::is_whitespace;
 
 #[derive(Clone, Copy, Eq, PartialEq, Debug)]
 pub(crate) enum StringType {
@@ -16,16 +17,35 @@ pub(crate) enum StringType {
 
 /// Parse a string that terminates based on some character(s) determined by `string_type`.
 pub(crate) fn parse_string(iter: &mut Iter<'_>, string_type: StringType) -> Result<String> {
+    let mut space_buffer = None;
+    let mut is_non_white_reached = false;
     let mut result = String::new();
     while !is_end_char(iter, string_type) {
         if iter.st.c == '&' {
             let c = parse_escape(iter)?;
-            result.push(c);
+            if is_whitespace(c) {
+                if is_non_white_reached {
+                    space_buffer = Some(' ');
+                }
+            } else {
+                is_non_white_reached = true;
+                result.push(c);
+            }
         } else if is_forbidden(iter, string_type) {
             return parse_err!(iter, "forbidden character in {:?} string", string_type);
+        } else if iter.is_whitespace() {
+            if is_non_white_reached {
+                space_buffer = Some(' ');
+            }
         } else {
+            is_non_white_reached = true;
+            if let Some(space) = space_buffer {
+                result.push(space);
+                space_buffer = None;
+            }
             result.push(iter.st.c);
         }
+
         if !iter.advance() {
             return parse_err!(
                 iter,