+++ testgen/xmltestgen/src/main/java/com/matthewjamesbriggs/xmltestgen/XNamespaces.java
@@ -1,0 +1,13 @@
+package com.matthewjamesbriggs.xmltestgen;
+
+public enum XNamespaces {
+    ON, OFF;
+
+    public static XNamespaces fromBoolean(boolean isNamespacesOn) {
+        if (isNamespacesOn) {
+            return ON;
+        } else {
+            return OFF;
+        }
+    }
+}